dog.h

+++ 0x0E-structures_typedef/dog.h
@@ -12,7 +12,7 @@ struct dog
 {
 	char *name;
 	char *owner;
-	int age;
+	float age;
 };
 
 /**
